comentarios

+++ proyectoIntegrador/controllers/productController.js
@@ -2,6 +2,7 @@
 
 const db = require('../database/models');
 const Producto = db.Producto;
+const Comentario = db.Comentario;
 // const Usuario = db.Usuario;
 
 const productController = {
@@ -48,16 +49,27 @@ const productController = {
             include: [{association:'usuarios'}, {association:'comentarios', include:[{association:'usuarios'}]}]})
         .then(function (data) {
             if(req.session.idUser == data.idUsuario){
-                Producto.destroy({
-                    where: {id : data.id}
-                })
-                .then(function(data){
-                    res.redirect('/')
-                })
-                .catch(function(err){
-                    console.log(err);
-                 })
-            } else {
+                if(data.comentarios != null){
+                    for(i=0; i<data.comentarios.length; i++){
+                        let idComentario = data.comentarios[i].id
+                        console.log(idComentario)
+                        Comentario.destroy({
+                            where: {id : idComentario}})
+
+                    }}
+                    Producto.destroy({
+                        where: {id : data.id}
+                    })
+                    .then(function(data){
+                        res.redirect('/')
+                    })
+                    .catch(function(err){
+                        console.log(err);
+                     })
+                
+
+                }     
+            else {
                 errors.message = 'Este producto no le pertenece, no puede eliminarlo'; //cargamos el mensaje
                 res.locals.errors = errors; //Usamos locals para pasarlo a la vista
                 res.render('product', {data:data}); //Renderizamos la vista 
@@ -107,7 +119,31 @@ const productController = {
     },
 
     comentario: function(req, res){
-
+        let errors = {}
+        let id = req.body.id
+        Producto.findByPk(id, {include:[{association:'comentarios', include: {association: 'usuarios'}}, {association: 'usuarios'}]})
+        .then(function(data){
+            if(req.session.user != null){
+                comment = req.body.comment
+                Comentario.create({
+                    idPost: data.id,
+                    idUsuario: req.session.idUser,
+                    comentario: comment,
+                    })
+                .then(function(info){
+                    res.redirect('/product/idProduct/' + id);  
+                })
+                .catch(function(err){
+                    console.log(err);
+                }) 
+            }
+            else{
+                res.redirect('/users/login')
+            }
+        })
+        .catch(function(err){
+            console.log(err);
+        }) 
     }
 }
